requirements.txt
# Django и основные зависимости
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0

# Для тестирования
pytest>=7.4.0
pytest-django>=4.5.0

# Для продакшена
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
